+++ src/bika/lims/content/client.py
@@ -120,7 +120,6 @@ schema = Organisation.schema.copy() + Schema((
         ),
     ),
 
-    # TODO Fix Client restricted categories are not considered in Add sample
     UIDReferenceField(
         "RestrictedCategories",
         schemata="Preferences",